Big code refactor
