+++ src/diff.rs
@@ -22,8 +22,8 @@ pub struct ChangedPublicItem {
 impl ChangedPublicItem {
     pub fn changed_tokens(&self) -> Vec<ChangedTokenStream> {
         ChangedTokenStream::new(ChangedPublicItem::align_tokens(
-            self.old.tokens(),
-            self.new.tokens(),
+            &self.old.tokens,
+            &self.new.tokens,
         ))
     }
 
@@ -194,7 +194,7 @@ impl PublicItemsDiff {
                     added.push(new);
                 }
                 (Some(old), Some(new)) => {
-                    if old != new && old.0.path == new.0.path {
+                    if old != new && old.path == new.path {
                         // The same item, but there has been a change in type
                         changed.push(ChangedPublicItem { old, new });
                     } else {
@@ -239,8 +239,6 @@ impl PublicItemsDiff {
 
 #[cfg(test)]
 mod tests {
-    use crate::item_iterator::PublicItemInner;
-
     use super::*;
 
     #[test]
@@ -342,11 +340,9 @@ mod tests {
     }
 
     fn item_with_path(path: &str) -> PublicItem {
-        PublicItem(PublicItemInner {
-            prefix: String::from("prefix "),
-            path: String::from(path),
-            suffix: String::from(" suffix"),
+        PublicItem {
+            path: path.split("::").map(|i| i.to_string()).collect(),
             tokens: TokenStream::default(),
-        })
+        }
     }
 }
