+++ src/lib.rs
@@ -32,9 +32,15 @@
 mod error;
 mod intermediate_public_item;
 mod item_iterator;
+mod render;
 pub mod tokens;
 
 pub mod diff;
+
+// Documented at the definition site so cargo doc picks it up
+pub use error::{Error, Result};
+
+pub use item_iterator::PublicItem;
 
 /// This constant defines the minimum version of nightly that is required in
 /// order for the rustdoc JSON output to be parsable by this library. Note that
@@ -45,34 +51,6 @@ pub mod diff;
 /// this library to support the latest format. If you use this version of
 /// nightly or later, you should be fine.
 pub const MINIMUM_RUSTDOC_JSON_VERSION: &str = "nightly-2022-03-14";
-
-// Documented at the definition site so cargo doc picks it up
-pub use error::Error;
-
-// Documented at the definition site so cargo doc picks it up
-pub use error::Result;
-
-/// Represent a public item of an analyzed crate, i.e. an item that forms part
-/// of the public API of a crate. Implements `Display` so it can be printed. It
-/// also implements [`Ord`], but how items are ordered are not stable yet, and
-/// will change in later versions.
-#[derive(Clone, PartialEq, Eq, PartialOrd, Ord)]
-pub struct PublicItem(item_iterator::PublicItemInner);
-
-impl PublicItem {
-    pub fn tokens(&self) -> &tokens::TokenStream {
-        &self.0.tokens
-    }
-}
-
-/// One of the basic uses cases is printing a sorted `Vec` of `PublicItem`s. So
-/// we implement `Display` for it.
-impl std::fmt::Display for PublicItem {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        //self.0.tokenize();
-        write!(f, "{}", self.0)
-    }
-}
 
 /// Contains various options that you can pass to [`public_items_from_rustdoc_json_str`].
 #[derive(Copy, Clone, Debug)]
@@ -143,11 +121,11 @@ pub fn public_items_from_rustdoc_json_str(
 ) -> Result<Vec<PublicItem>> {
     let crate_: rustdoc_types::Crate = serde_json::from_str(rustdoc_json_str)?;
 
-    let public_items: Vec<_> = item_iterator::public_items_in_crate(&crate_, options).collect();
+    let mut public_items: Vec<_> = item_iterator::public_items_in_crate(&crate_, options).collect();
 
-    //if options.sorted {
-    //    public_items.sort();
-    //}
+    if options.sorted {
+        public_items.sort();
+    }
 
     Ok(public_items)
 }
